requirements.txt
joblib=1.2.0
anyio==3.3.2
captcha-solver==0.1.5
certifi==2021.5.30
charset-normalizer==2.0.6
dacite==1.6.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
numpy==1.21.2
onlinesimru==1.0.17
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
urllib3==1.26.7
random-password-generator==2.2.0
